requirements.txt
google-cloud-bigquery==1.23.1